+++ src/byob.js
@@ -108,7 +108,7 @@ WatcherMorph, XML_Serializer, SnapTranslator, SnapExtensions*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.byob = '2021-October-12';
+modules.byob = '2021-October-14';
 
 // Declarations
 
@@ -1753,6 +1753,12 @@ BlockDialogMorph.prototype.init = function (target, action, environment) {
     this.categories = new BoxMorph();
     this.categories.color = SpriteMorph.prototype.paletteColor.lighter(8);
     this.categories.borderColor = this.categories.color.lighter(40);
+    this.categories.buttons = [];
+
+    this.categories.refresh = function () {
+        this.buttons.forEach(cat => cat.refresh());
+    };
+
     this.createCategoryButtons();
     this.fixCategoriesLayout();
     this.add(this.categories);
@@ -1775,9 +1781,7 @@ BlockDialogMorph.prototype.openForChange = function (
     this.category = category;
     this.blockType = type;
 
-    this.categories.children.forEach(each =>
-        each.refresh()
-    );
+    this.categories.refresh();
     this.types.children.forEach(each => {
         each.setColor(clr);
         each.refresh();
@@ -1833,9 +1837,7 @@ BlockDialogMorph.prototype.addCategoryButton = function (category) {
         this, // this block dialog box is the target
         () => {
             this.category = category;
-            this.categories.children.forEach(each =>
-                each.refresh()
-            );
+            this.categories.refresh();
             if (this.types) {
                 this.types.children.forEach(each =>
                     each.setColor(colors[2])
@@ -1863,6 +1865,7 @@ BlockDialogMorph.prototype.addCategoryButton = function (category) {
     button.fixLayout();
     button.refresh();
     this.categories.add(button);
+    this.categories.buttons.push(button);
     return button;
 };
 
@@ -1882,9 +1885,7 @@ BlockDialogMorph.prototype.addCustomCategoryButton = function (category, clr) {
         this, // this block dialog box is the target
         () => {
             this.category = category;
-            this.categories.children.forEach(each =>
-                each.refresh()
-            );
+            this.categories.refresh();
             if (this.types) {
                 this.types.children.forEach(each =>
                     each.setColor(colors[2])
@@ -1912,6 +1913,7 @@ BlockDialogMorph.prototype.addCustomCategoryButton = function (category, clr) {
     button.fixLayout();
     button.refresh();
     this.categories.add(button);
+    this.categories.buttons.push(button);
     return button;
 };
 
@@ -1924,8 +1926,14 @@ BlockDialogMorph.prototype.fixCategoriesLayout = function () {
         border = 10, // this.categories.border,
         l = this.categories.left(),
         t = this.categories.top(),
+        scroller,
         row,
-        col;
+        col,
+        i;
+
+    this.categories.setWidth(
+        3 * xPadding + 2 * buttonWidth
+    );
 
     this.categories.children.forEach((button, i) => {
         if (i < 8) {
@@ -1950,17 +1958,43 @@ BlockDialogMorph.prototype.fixCategoriesLayout = function () {
         this.categories.border = 0;
         this.categories.edge = 0;
     }
-    this.categories.setWidth(
-        3 * xPadding + 2 * buttonWidth
-    );
 
-    this.categories.setHeight(
-        (5 + 1) * yPadding
-            + 5 * buttonHeight
-            + (more ? (more * (yPadding + buttonHeight) + border / 2) : 0)
-            + 2 * border
-    );
+    if (more > 6) {
+        scroller = new ScrollFrameMorph(
+            null,
+            null,
+            SpriteMorph.prototype.sliderColor.lighter()
+        );
+        scroller.setColor(this.categories.color);
+        scroller.acceptsDrops = false;
+        scroller.contents.acceptsDrops = false;
+        scroller.setPosition(
+            new Point(
+                this.categories.left() + this.categories.border,
+                this.categories.children[10].top()
+            )
+        );
+        scroller.setWidth(this.categories.width() - this.categories.border * 2);
+        scroller.setHeight(buttonHeight * 6 + yPadding * 5);
 
+        for (i = 0; i < more; i += 1) {
+            scroller.addContents(this.categories.children[10]);
+        }
+        this.categories.add(scroller);
+        this.categories.setHeight(
+            (5 + 1) * yPadding
+                + 5 * buttonHeight
+                + 6 * (yPadding + buttonHeight) + border + 2
+                + 2 * border
+        );
+    } else {
+        this.categories.setHeight(
+            (5 + 1) * yPadding
+                + 5 * buttonHeight
+                + (more ? (more * (yPadding + buttonHeight) + border / 2) : 0)
+                + 2 * border
+        );
+    }
 };
 
 // type radio buttons
@@ -4636,7 +4670,7 @@ BlockVisibilityDialogMorph.prototype.hideBlocks = function () {
     }
     ide.flushBlocksCache();
     ide.refreshPalette();
-    ide.recordUnsavedChanges()
+    ide.recordUnsavedChanges();
 };
 
 // BlockVisibilityDialogMorph layout