+++ src/objects.js
@@ -87,7 +87,7 @@ BlockVisibilityDialogMorph*/
 
 /*jshint esversion: 6*/
 
-modules.objects = '2021-October-12';
+modules.objects = '2021-October-14';
 
 var SpriteMorph;
 var StageMorph;
@@ -2955,7 +2955,7 @@ SpriteMorph.prototype.makeBlock = function () {
     );
     if (category !== 'variables' || category !== 'unified') {
         dlg.category = category;
-        dlg.categories.children.forEach(each => each.refresh());
+        dlg.categories.refresh();
         dlg.types.children.forEach(each => {
             each.setColor(clr);
             each.refresh();